Upgrade presto-hyperloglog to support 0.177
Along with updating the code to work with renamed / removed libraries
this also makes a few additional changes.

* Switches the packaging to use presto-plugin instead of an uberjar
* Removes compile-time dependencies on the presto-main library
* Upgrades algebird to 0.13.0 (mainly to keep track of upstream)

The most important of those is removing the compile-time dependencies on
presto-main as the presto-spi interface is considered more stable and
should require fewer code changes when upgrading the library to support
newer versions of presto-server.